[Fix] planterior 데이터 클래스 구분
