+++ lib/list/planteriordata.dart
@@ -1,0 +1,63 @@
+class PlanteriorData {
+  final List<String> images;
+  final List<String> contents;
+  final List<String> hashtags;
+  final List<String> dates;
+
+  PlanteriorData(
+      {required this.images,
+      required this.contents,
+      required this.hashtags,
+      required this.dates});
+
+  static PlanteriorData get defaultData => PlanteriorData(
+        images: [
+          'assets/planterior/planterior1.png',
+          'assets/planterior/planterior2.png',
+          'assets/planterior/planterior3.png',
+          'assets/planterior/planterior4.png',
+          'assets/planterior/planterior5.png',
+          'assets/planterior/planterior6.png',
+          'assets/planterior/planterior7.png',
+          'assets/planterior/planterior8.png',
+          'assets/planterior/planterior9.png',
+          'assets/planterior/planterior10.png',
+        ],
+        contents: [
+          '[오늘의 플랜테리어🍀]\n식물과 함께하는 인테리어 어떠세요? \n침실 옆에 식물을 두고 싱그러운 아침을 맞이하세요!',
+          '[오늘의 플랜테리어🍀]\n이제는 나도 식집사! \n저희 집 욕실 어떤가요?🛁\n싱그러운 욕실과 만나보아요 :)',
+          '[오늘의 플랜테리어🍀]\n공기 정화에 탁월한 스타피필름, 화장실 앞에 둬보세요!\n향기로운 공간 되세요 :)',
+          '[오늘의 플랜테리어🍀]\n식물과 함께하는 주말 어떠신가요?',
+          '[오늘의 플랜테리어🍀]\n활력과 생기를 불어 넣어주는 식물\n원목 소재의 가구와 함께 배치해보아요 :)',
+          '[오늘의 플랜테리어🍀]\n화사한 분위기의 방\n식물로 한층 더 화사해져요 :)',
+          '[오늘의 플랜테리어🍀]\n실패하지 않는 플랜테리어 조합\n화이트톤의 방과 우드톤의 화분 어떠신가요?\n작은 식물로 방 분위기를 바꿔봐요 :)',
+          '[오늘의 플랜테리어🍀]\n심심한 주방, 식탁 위에 식물을 둬보세요!\n식용 허브를 둔다면 활용도 만점!!',
+          '[오늘의 플랜테리어🍀]\n나란히 두는 화분\n채광 좋은 집에서만 할 수 있는 특권, 누려보세요!!',
+          '[오늘의 플랜테리어🍀]\n오늘부터 우리집의 허전한 공간을 채워줄\n식물을 한번 찾아보시면 어떨까요?',
+        ],
+        hashtags: [
+          '#플랜테리어 #아레카야자 #식물 #초보자가_키우기_쉬운_식물',
+          '#식집사 #식물 #공기정화_식물',
+          '#플랜테리어 #식물 #초보자가_키우기_쉬운_식물',
+          '#식집사 #식물 #공기정화_식물',
+          '#플랜테리어 #식물 #초보자가_키우기_쉬운_식물',
+          '#식집사 #식물 #공기정화_식물',
+          '#플랜테리어 #식물 #초보자가_키우기_쉬운_식물',
+          '#식집사 #식물 #공기정화_식물',
+          '#플랜테리어 #식물 #초보자가_키우기_쉬운_식물',
+          '#식집사 #보스턴고사리 #식물 #공기정화_식물',
+        ],
+        dates: [
+          '2024년 03월 01일',
+          '2024년 03월 02일',
+          '2024년 03월 03일',
+          '2024년 03월 04일',
+          '2024년 03월 05일',
+          '2024년 03월 06일',
+          '2024년 03월 07일',
+          '2024년 03월 08일',
+          '2024년 03월 09일',
+          '2024년 03월 10일',
+        ],
+      );
+}
