+++ lib/mypage/purchase.dart
@@ -1,12 +1,9 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:greentouch/layout/appbar.dart';
 import 'package:greentouch/layout/appbar_back.dart';
 import 'package:greentouch/list/purchase_complete.dart';
-import 'package:greentouch/mypage/finished.dart';
 import 'package:greentouch/service/cart_service.dart';
 import 'package:provider/provider.dart';
-import '../layout/app_drawer.dart';
+
 import '../product/plant_service.dart';
 
 class PurchasePage extends StatefulWidget {
